+++ src/components/NumberBox/index.ts
@@ -1,0 +1,15 @@
+export { NumberBox } from './NumberBox';
+
+export type {
+  NumberBoxProps,
+  NumberBoxVariant,
+  NumberBoxSize,
+  NumberBoxLabelPosition,
+  NumberBoxTheme,
+  NumberBoxVariantTheme,
+  NumberBoxThemePreset,
+  NumberBoxThemeInput,
+  NumberBoxOnChangeHandler,
+  NumberBoxOnFocusHandler,
+  NumberBoxOnBlurHandler,
+} from './type/NumberBox.types';